Refactor typography components
