+++ packages/ui/src/components/modal.tsx
@@ -1,6 +1,6 @@
 import React, { ReactNode } from 'react'
 import styled, { ThemedStyledProps } from 'styled-components'
-import { Text } from './page/Typography/Text'
+import { Text } from './typography'
 import { BorderRad, Colors, Shadows } from '../constants'
 import { Close } from './buttons/CloseButton'
 import { CrossIcon } from './icons/CrossIcon'
