+++ packages/ui/src/components/selects/AccountSelectTemplate/OptionAccount.tsx
@@ -5,7 +5,7 @@ import { Account } from '../../../hooks/types'
 import { useBalance } from '../../../hooks/useBalance'
 import { BalanceInfo, InfoTitle, InfoValue } from '../../../modals/common'
 import { AccountInfo } from '../../AccountInfo'
-import { TokenValue } from '../../page/Typography/JoyValue'
+import { TokenValue } from '../../typography'
 
 export interface SelectAccountOption {
   account: Account
