+++ packages/ui/src/components/typography/Label.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components'
-import { Colors } from '../../../constants'
+import { Colors } from '../../constants'
 
 export const Label = styled.span`
   display: inline-grid;
