+++ packages/ui/src/components/typography/index.ts
@@ -1,0 +1,3 @@
+export * from './Label'
+export * from './Text'
+export * from './TokenValue'
