+++ packages/ui/src/modals/TransferModal/SignTransferModal.tsx
@@ -8,7 +8,7 @@ import { ButtonPrimaryMedium } from '../../components/buttons/Buttons'
 import { ArrowDownExpandedIcon } from '../../components/icons/ArrowDownExpandedIcon'
 import { Modal, ModalBody, ModalFooter, ModalHeader, SignTransferContainer } from '../../components/modal'
 import { HelpNotification } from '../../components/notifications/HelpNotification'
-import { TokenValue } from '../../components/page/Typography/JoyValue'
+import { TokenValue } from '../../components/typography'
 import { Account } from '../../hooks/types'
 import { useApi } from '../../hooks/useApi'
 import { useBalance } from '../../hooks/useBalance'
