+++ packages/ui/src/modals/TransferModal/TransactionSuccessModal.tsx
@@ -4,7 +4,7 @@ import { AccountInfo } from '../../components/AccountInfo'
 import { ArrowDownExpandedIcon } from '../../components/icons/ArrowDownExpandedIcon'
 import { SuccessIcon } from '../../components/icons/SuccessIcon'
 import { ModalHeader, ResultModal, SuccessModalBody } from '../../components/modal'
-import { TokenValue } from '../../components/page/Typography/JoyValue'
+import { TokenValue } from '../../components/typography'
 import { Account } from '../../hooks/types'
 import { useBalance } from '../../hooks/useBalance'
 import {
