+++ packages/ui/src/modals/TransferModal/TransferDetailsModal.tsx
@@ -4,7 +4,7 @@ import styled from 'styled-components'
 import { AccountInfo } from '../../components/AccountInfo'
 import { ButtonPrimaryMedium, ButtonSecondarySmall } from '../../components/buttons/Buttons'
 import { Modal, ModalBody, ModalFooter, ModalHeader } from '../../components/modal'
-import { TokenValue } from '../../components/page/Typography/JoyValue'
+import { TokenValue } from '../../components/typography'
 import { SelectAccount } from '../../components/selects/AccountSelectTemplate/SelectAccount'
 import { BorderRad, Colors } from '../../constants'
 import { Account } from '../../hooks/types'
