+++ packages/ui/src/pages/Profile/Accounts.tsx
@@ -6,7 +6,7 @@ import { ButtonGhostMediumSquare } from '../../components/buttons/Buttons'
 import { ArrowDownIcon } from '../../components/icons/ArrowDownIcon'
 import { PageTab, PageTabs } from '../../components/page/PageTabs'
 import { PageTitle } from '../../components/page/PageTitle'
-import { TokenValue } from '../../components/page/Typography/JoyValue'
+import { TokenValue } from '../../components/typography'
 import { TransferButton } from '../../components/TransferButton'
 import { BorderRad, Colors, Sizes } from '../../constants'
 import { Account } from '../../hooks/types'
