+++ packages/ui/src/pages/Profile/TotalBalances.tsx
@@ -3,8 +3,8 @@ import styled from 'styled-components'
 import BN from 'bn.js'
 import { ButtonPrimarySmall } from '../../components/buttons/Buttons'
 import { HelpNotification } from '../../components/notifications/HelpNotification'
-import { Label } from '../../components/page/Typography/Label'
-import { TokenValue } from '../../components/page/Typography/JoyValue'
+import { Label } from '../../components/typography'
+import { TokenValue } from '../../components/typography'
 import { BorderRad, Colors, Shadows } from '../../constants'
 import { useTotalBalances } from '../../hooks/useTotalBalances'
